% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secenv.R
\name{load_secret_Renviron}
\alias{load_secret_Renviron}
\title{Load GPG-protected environment file}
\usage{
load_secret_Renviron(fname = "~/.Renviron-secret")
}
\arguments{
\item{fname}{file name to read}
}
\value{
Nothing
}
\description{
Load GPG-protected environment file
}

