% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secenv.R
\name{read_Renviron}
\alias{read_Renviron}
\title{Parse Renviron file}
\usage{
read_Renviron(fname)
}
\arguments{
\item{fname}{character, file to parse}
}
\value{
Named list of environment variables
}
\description{
Parse Renviron file
}

