% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secenv.R
\name{write_Renviron}
\alias{write_Renviron}
\title{Write environment variables to file}
\usage{
write_Renviron(l, output, ...)
}
\arguments{
\item{l}{named list of environment variables}

\item{output}{connection or character string where to write to, passed to \code{\link{writeLines}}}

\item{...}{other arguments passed to \code{\link{writeLines}}}
}
\value{
Nothing
}
\description{
Write environment variables to file
}

